==> 3_etl.ipynb <==
# ETL

#!which python

import pandas as pd
import sqlite3
import numpy as np
import re
import pycountry_convert as pc

## Import the data from our database

db = 'dataBases//world_cups_squads.db'

def load_table_as_df(database):
    conn = sqlite3.connect(database)
    df = pd.read_sql_query("SELECT * FROM squads", conn)
    conn.close()
    return df

df = load_table_as_df(db)

## Explore the data

df.head()

df.info()

### Search for null values and treat them if needed
    

df.isna().any().any()

def get_nan_in_df(dataFrame):
    # count all null values in each column, sort them by the amount of null values.
    nanValues = dataFrame.isna().sum().sort_values(ascending=False)
    # get the percentage of those values
    nanValuesPrct = (nanValues/len(dataFrame) )*100
    
    return nanValuesPrct

get_nan_in_df(df)

## Data Transformation

We'll drop the following columns:


|Column        |Reason                            |
|--------------|----------------------------------|
|    "goals"   |86% of the column values are null.|
|"shirt_number"| It's irrelevant for our analysis |


df.drop(columns = ['goals','shirt_number'], inplace = True)

### convert 'wc_date' data type to datetime

df['wc_date'] = pd.to_datetime(df['wc_date'])

### Create a column to contain the year of a given world cup

df['year'] = df['wc_date'].dt.strftime('%Y').astype(int)

df.info()

capsNan = df[df.caps.isna()]
len(capsNan)

Even though there are only a few nan registers, we'd like to keep them all and see if we can get the data from other sources or input a value. 

 We'll calculate the median among these footballer's teammates. Then we'll apply it to the formers.

#### Optimize memory usage

# 16 bits range goes from −32,768 to 32,767

df.year = pd.Series(df.year, dtype=pd.Int16Dtype()) #from now on it'll use pandas.NA as missing value, instead of numpy.nan.        

### replace

Let's find the different values in the column 'caps'

# get a data frame that contains both, the unique values and their counts 
value_caps = df.caps.value_counts().reset_index().rename(columns={'index':'value'})

# convert the values to a list
values_caps_list = value_caps.value.tolist()

#values_caps_list.sort()  # can't sort this time, because the list has a mix of str and int values
#values_caps_list

# Let's see if we can filter the weird values. They don't look like integers, so let's try the following.
for element in values_caps_list:
    if type(element) ==str:
        print(element)

# Great! Let's replace them.
df.caps.replace({'?':np.nan, '-': np.nan, '0-': 0, '1[7]':1, '7*':7, '32*':32  }, inplace=True ) 

We've added a few nan values. So let's check the missing values count again.

capsNan = df[df.caps.isna()]
len(capsNan)

### Create a list of tuples with all the teams having nan values in column 'caps'

nan_grouped = capsNan.groupby(['national_team','year'])['caps'].unique()

nan_grouped = nan_grouped.to_frame().reset_index().drop(columns = 'caps')
#nan_grouped

unique_nan_list = list(zip(nan_grouped.national_team, nan_grouped.year))

print(unique_nan_list)

Remove the (1950, Paraguay) tuple, since all their corresponding cells in the column 'caps' are null.

unique_nan_list.remove(('Paraguay', 1950))

### Create a column to register the imputed values

df['imputed']  = '-'

### Impute values

def get_a_team(nation_name, wc_year):
    team = df[(df.national_team == nation_name) &(df.year == wc_year)]
    return team    

def get_team_median(nation_name, wc_year):
    # Obtain a data frame of a particular team 
    team = get_a_team(nation_name, wc_year)
    # and get its caps median
    team_median = team.caps.median()
    # In case the median is a rational number, round it up
    team_median = np.ceil(team_median)
    return team_median

def get_team_nans_list(nation_name, wc_year, column_name):
    # call get_a_team to get the data frame of the desired team
    team_nans = get_a_team(nation_name, wc_year)
    # filter the data frame by nan values in the selected column(passed as parameter)
    team_nans = team_nans[(team_nans[column_name].isna()== True)]
    # make a list with the indexes of the filtered data frame
    listed_by_index = team_nans.index.to_list()
    return listed_by_index

# Void function
def fill_nan_in_team(nation_name, wc_year, column_name):
    # Get all the indexes that represent footballers with no caps data, and for each of them 
    for index in get_team_nans_list(nation_name, wc_year, column_name):
    # calculate the respective team's median. Then, impute this median to each record in the original data frame.
        df.at[index, column_name] = get_team_median(nation_name, wc_year)        

# This function records and points out in which column of the data frame an imputation was made.
def record_imputation(a_list,column_name):
    for element in a_list:
        df.at[element,'imputed'] = column_name        

def fill_all(list_of_tuples, name_of_column):
    # iterate over a list of tuples (of the form (country, year)) 
    for elemento in list_of_tuples:
        # call the functions above to do all the job at once
        list_of_indexes = get_team_nans_list(elemento[0], elemento[1], name_of_column)
        fill_nan_in_team(elemento[0], elemento[1], name_of_column)
        record_imputation(list_of_indexes, name_of_column)

fill_all(unique_nan_list, 'caps')

Let's fill in the  null values of  the Paraguay national team of 1950.

capsNan = df[df.caps.isna()]
len(capsNan)

#capsNan

list_indexes = capsNan.index.to_list()

print(list_indexes)

We decided to calculate the median of all the South American teams in the tournament and impute that value to the Paraguayan national team.

df1950 = df[df.year == 1950]

df1950Bol = df1950[(df1950.national_team == 'Bolivia')]
df1950Bra = df1950[(df1950.national_team == 'Brazil')]
df1950Chi = df1950[(df1950.national_team == 'Chile')]
df1950Uru = df1950[(df1950.national_team == 'Uruguay')]

df1950SAm = pd.concat([df1950Bol, df1950Bra, df1950Chi, df1950Uru])

for element in list_indexes:
        # impute value
        df.at[element,'caps']  = df1950SAm.caps.median()
        # record the imputation
        df.at[element,'imputed'] = 'caps'

# optimize the memory usage
df.caps = pd.Series(df.caps, dtype=pd.Int16Dtype())

### Column 'national_team'

Normalize Yugoslavia's name.

teams = df.national_team.unique().tolist()
teams.sort()

print(teams)

yugo = df[df.national_team == 'FR Yugoslavia'].index.tolist()

for ind in yugo:
    df.at[ind, 'national_team'] = 'Yugoslavia'

### Column 'club_origin'

club_originNan = df[df.club_origin.isna()]
club_originNan

df.at[3303,'club'], df.at[3303,'club_origin'] = 'Free Agent', '-'
df.at[7973,'club'], df.at[7973,'club_origin'] = 'Kavala', 'Greece'
df.at[8345,'club'], df.at[8345,'club_origin'] = 'San Jose Earthquakes', 'United States'
df.at[8355,'club'], df.at[8355,'club_origin'] = 'Wellington Phoenix', 'New Zealand'

# normalize Yugoslavia's name
#df['club_origin'] = df[df['club_origin'].apply(lambda x: x.replace('Socialist Federal Republic of',''))]

# get_nan_in_df(df)

### Normalize column 'player'

As far as we can see, this data has several punctuation marks we need to remove.

punctuation = ['!','"','#','$','~','%','&','\'','(',')','*','+',',','-','.','/',':',';',
               '<','=','>','?','@','[','\\',']','^','_','`','{','|','}']

We can filter the data frame by all the registers in the column 'player' that contain punctuation marks.

def find_punctuation(a_string):
    for element in punctuation:
        if element != '-':
            if element in a_string:
                return True
    return False            

dfPunt = df[df.player.apply(find_punctuation)]

#dfPunt

# Split the strings on the first unwanted char, then keep the first element of the resulting list, that's the desired value.
df.player = df.player.apply(lambda x: x.split('[')[0])
df.player = df.player.apply(lambda x: x.split(' (')[0])
df.player = df.player.apply(lambda x: x.split('*')[0])

# Getting rid of footballers' nicknames by removing all those chars between quotation marks or between apostrophes
df.player = df.player.apply(lambda x: re.sub(r'"([A-Za-z0-9_\./\\-]*)"' ,'' , x))
df.player = df.player.apply(lambda x: re.sub(r"'([A-Za-z0-9_\./\\-]*)'" ,'' , x))

from unidecode import unidecode

# Apply unidecode to column 'player' to normalize the name of the players
df['player'] = df['player'].apply(lambda x: unidecode(x))

df

### Normalize  column 'birthday'

list_of_birthdays = df.birthday.tolist()

list_of_birthdays

After looking these values for a while, we've found a couple of patterns.
1) (d-d-d) where d stands for digit
2) d w d (w d) where w stands for letters and d for digits
3) d\n     Again, d is a digit and \n is the new line char
4) \n      a mere new line char

# Remove all whitespaces to avoid losing data while using reverse split with whitespaces as separators.
df.birthday = df.birthday.apply(lambda x: x.strip())   

# Remove all the chars to the right of the year digit in the data with the second pattern type.
df.birthday = df.birthday.apply(lambda x: x.rsplit(' ', maxsplit = 2)[0])

# remove the parentheses
df.birthday = df.birthday.apply(lambda x: x.replace('(','').replace(')',''))

For data containing only one year, We'll impute the median date of that year, that's July 2.

indices = df[(df.birthday.str.len() == 4)].index.tolist()  # filter by those cells having only one year value

def complete_birthday(bd_string):
    if len(bd_string) == 4:
        bd_string = f'{bd_string}-07-02'
         
    return bd_string

df.birthday = df.birthday.apply(lambda x: complete_birthday(x))

We've imputed some birthdays, now we'll record this imputation in the corresponding column.

for index in indices:
    old = df.at[index,'imputed']
    if old == '-':
        old =  old.replace('-', '')
    df.at[index,'imputed'] = old + ' birthday'

Now let's change the data type in this column, this should be final the step for this field.

df['birthday'] = pd.to_datetime(df['birthday'])

df.birthday.isna().sum()

There are still 40 null birthday values, but we are good to go because we need this field to calculate players' age. So, we can take a central tendency measure for their age, later.

## Normalize column club

# Aplicar unidecode a la columna 'club' y crear una nueva columna
df['club'] = df['club'].apply(lambda x: unidecode(x))

# Utilizar el método .str.replace() para realizar la sustitución.Es lo mismo que hacerlo con apply + función anónima pero 
# más eficiente.
df['club'] = df['club'].str.replace(r'\[[^\]]*\]', '', regex=True)

df['club'] = df['club'].str.replace('1. FC ', '', regex = False)
df['club'] = df['club'].str.replace('1. ', '', regex = False)

df['club'] = df['club'].str.replace(r'(^|\s)(AC|A\.C\.)(\s|$)', '', regex=True)

df['club'] = df['club'].str.replace(r'(^|\s)(CD|C\.D\.)(\s|$)', '', regex=True)

df['club'] = df['club'].str.replace(r'(^|\s)(CF|C\.F\.)(\s|$)', '', regex=True)

df['club'] = df['club'].str.replace(r'(^|\s)(Club)(\s|$)', '', regex=True)

df['club'] = df['club'].str.replace(r'(^|\s)(FC|F\.C\.)(\s|$)', '', regex=True)

df['club'] = df['club'].str.replace(r'(^|\s)(SC|S\.C\.)(\s|$)', '', regex=True)

df['club'] = df['club'].str.replace(r'(^|\s)(SK|S\.K\.)(\s|$)', '', regex=True)

df['club'] = df['club'].str.replace(r'(^|\s)(AS|A\.S\.)(\s|$)', '', regex=True)

df['club'] = df['club'].str.replace('Al-', 'Al ', regex = False)


normalized_names = { 'Admira/Wacker': 'Admira Wacker', 'Ajax': 'Ajax Amsterdam', 'Atletico Bilbao': 'Athletic Bilbao',
                    'Atlético Bilbao': 'Athletic Bilbao', 'Atletico de Bilbao': 'Athletic Bilbao', 
                    'Austria Vienna': 'Austria Wien', 'BV Borussia Dortmund': 'Borussia Dortmund',
                    'Bayern Munich': 'Bayern Munchen', "Beijing Guo'an": 'Beijing Guoan', 'Besiktas': 'Besiktas J. K.', 
                    'Bournemouth': 'AFC Bournemouth', 'CSKA SZ Sofia': 'CSKA Sofia', 'Colo Colo': 'Colo-Colo', 
                    'Cotonsport Garoua': 'Cottonsport Garoua',    'DFS Lokomotiv Plovdiv': 'DFS Lokomotiv Plovdiv',
                    'Dynamo Minsk,': 'Dinamo Minsk', 'Estudiantes La Plata': 'Estudiantes', 
                    'Estudiantes (LP)': 'Estudiantes', 'ExcelsiorRoubaix': 'Excelsior Roubaix',
                    'Ferencvaros': 'Ferencvarosi', 'Ferencvarosi TC': 'Ferencvarosi','Feijenoord': 'Feyenoord',
                    'Feyenoord Rotterdam': 'Feyenoord', 'Flamengo*': 'Flamengo', 'Free agent': 'Free Agent',
                    'GrasshopperZurich': 'Grasshopper Zurich', 'GrasshoppersZurich': 'Grasshopper Zurich',
                    'Internazionale': 'Inter Milan', 'Internazionale Milano': 'Inter Milan',
                    'Internazionale Milan': 'Inter Milan', 'Lille OSC': 'Lille', 'Malmo FF': 'Malmo',
                    'Moscow': 'CSKA Moscow', 'Lyon': 'Olympique Lyon', 'Marseille': 'Olympique Marseille',
                    'Olympique de Marseille': 'Olympique Marseille', 'PSV': 'PSV Eindhoven',
                    'Racingde France': 'Racing Paris', 'Rapid Vienna': 'Rapid Wien',
                    'Royal Olympicde Charleroi': 'Olympic Charleroi', 'Royal Racingde Bruxelles': 'Royal Racing Bruxelles',
                    'Royal Racingde Gand': 'Royal Racing Gand', 'Royal DaringMolenbeek': 'Royal Daring Molenbeek',
                    'Royal StandardLiege': 'Royal Standard Liege', 'RoyalBrugeois': 'Brugge KV', 'Brujas': 'Brugge KV',
                    'Royal Brugeois': 'Brugge KV', 'RoyalLiegeois': 'Liege', 'RoyalMalinois': 'Royal Malinois',
                    'SG Dynamo Dresden': 'Dynamo Dresden', 'Talleres de Cordoba': 'Talleres',
                    'Tatabanyai Banyasz': 'Tatabanyai Banyasz SE', 'Tecos UAG': 'Tecos', 'Tianjin TEDA': 'Tianjin Teda',
                    'Tigres UANL': 'Tigres', 'UANL': 'Tigres', 'UNAM Pumas': 'Pumas', 'UNAM': 'Pumas', 
                    'Union St. Gilloise': 'Union Royale Saint-Gilloise', 'White Star Brussel': 'White Star'                                            
                    }

# Aplicar los cambios en la columna
df['club'] = df['club'].replace(normalized_names)

#clubs = df.club.to_frame()
#clubs = clubs.club.drop_duplicates().to_frame()
#clubs = clubs.sort_values(by = 'club').reset_index(drop = True)

#clubs

### Create the  column 'age'

df.info()

df

df['age'] =  (df['wc_date'] - df['birthday']).dt.days //365

We want to fill these missing values with the median age of the world cup that the footballers with no age registered played in.

# This function receives an index list to point to the rows were we want to apply the changes
def fill_nan_age(indices_list, wc_age_median, column_name, year):
    for ind in indices_list:
        # If the year passed as a parameter matches the value in the column 'year' at the given index,
        if df.at[ind, 'year'] == year:
            # fill the missing value with the median of the corresponding world cup
            df.at[ind, column_name] = wc_age_median
            # then call another function to record the imputation
            record_age_imputation(ind, column_name)

def record_age_imputation(an_index, column_name):
    old = df.at[an_index, 'imputed'] 
    if old == '-':
        old =  old.replace('-', '')
    df.at[an_index, 'imputed'] = old + ' ' + column_name

# get all the nan values in the corresponding column
dfna= df[df.age.isna() == True]
#dfna

# make a list with the indices of the rows containing nan values in the column 'age' 
dfnaList = df[df.age.isna() == True].index.tolist()

# Get all the years of the World Cups corresponding to these missing values,
for year in dfna.year.unique():
    # and for each World Cup calculate the players' median age
    median = df[df['year'] == year].age.median()
    # fill the measing values with the calculated median
    fill_nan_age(dfnaList, median, 'age', year)

df.age = df.age.astype(int)

## Add a new 'confederation' column

We'll use pycountry to convert country names to continent names

# create the new column
df['confederation'] = ' '

# dictionary with all the countries names that pycountry couldn't procces
non_listed = {'China PR': 'Asia',
             'Czechoslovakia': 'Europe',
             'Dutch East Indies': 'Asia',
             'East Germany': 'Europe',
             'England':'Europe',
             'Northern Ireland':'Europe',
             'Republic of Ireland':'Europe',
             'Scotland': 'Europe',
             'Serbia and Montenegro': 'Europe',
             'Soviet Union': 'Europe',
             'Wales': 'Europe',
             'West Germany': 'Europe',
             'Yugoslavia': 'Europe',
             'Zaire': 'Africa'}

# this function takes a country name and returns the corresponding continent name
def country_to_continent(country_name):
    try:
        country_alpha2 = pc.country_name_to_country_alpha2(country_name)
        country_continent_code = pc.country_alpha2_to_continent_code(country_alpha2)
        country_continent_name = pc.convert_continent_code_to_continent_name(country_continent_code)
        
    except KeyError:
        country_continent_name =  non_listed.get(country_name)
    return country_continent_name

df['confederation'] = df.national_team.apply(lambda x: country_to_continent(x))

## Create a key column to identify players

df

# Create a key to identify the players. We'll get unique values later.
df['key'] = df['player'] + ' ' + df['birthday'].astype(str)

# set the column 'key' as the first one of our dataframe
df.insert(0, 'key', df.pop('key'))

df

## Rearrange the order of the columns

df = df[['key', 'player','national_team','position','birthday','age','caps',
         'club','club_origin','wc_date','year','confederation','imputed']]

## Save as csv

df.to_csv('dataSets\\allTimeSquads.csv', index = False)
