==> 4_maps.ipynb <==
import pandas as pd
import geopandas as gpd

# Load the shape file 
world = gpd.read_file('world_countries\\World_Countries.shp')
world = world.rename(columns={'COUNTRY': 'country' })

world

# Let's eliminate Antarctica
world.iloc[14]
world.drop(14, inplace=True)
world = world.reset_index(drop=True)

uk =  gpd.read_file('NUTS1\\NUTS1_Jan_2018_UGCB_in_the_UK.shp')

#uk

uk.drop(['nuts118cd', 'bng_e', 'bng_n', 'long', 'lat', 'GlobalID'], axis = 1, inplace = True)

uk = uk.rename(columns={'nuts118nm': 'country'}) 

uk.crs

world.crs

As we can see, our GeoDataFrames use different coordinate reference systems. So, we need to reconcile this data.

# Assign the crs value to uk GeoDataFrame
uk = uk.to_crs('EPSG:4326')

# Create one dataframe containing England regions
england = uk.iloc[:9]
# and another one for the other 3 UK countries
other = uk.iloc[-3:]

england

other

# Convert England regions dataframe to a  GeoDataFrame
eng_gdf = gpd.GeoDataFrame(england, geometry='geometry')

# save the GeoDataFrame
eng_gdf.to_file('england\\regions\\england_regions.shp')

england_regions =  gpd.read_file('england\\regions\\england_regions.shp')

england_regions

# join all the polygons to get the shape of England
england = england_regions.unary_union

# Create a new GeoDataFrame of England
england_gdf = gpd.GeoDataFrame({'country': 'England', 'geometry': [england]})

# We need to specify a coordinate reference system(CRS) for our new GeoDataFrame
england_gdf.crs = 'EPSG:4326'
# save it
england_gdf.to_file('england\\country\\england.shp')

# Add England to 'world' GeoDataFrame
world = pd.concat([world, england_gdf], ignore_index=True)

# Create a GeoDataFrame for Wales, Scotland and Nothern Ireland
other_gdf = gpd.GeoDataFrame(other, geometry='geometry')


# We need to specify a coordinate reference system(CRS) for our new GeoDataFrame
other_gdf.crs = 'EPSG:4326'


# save to a shapefile
other_gdf.to_file('rest_of_uk\\other.shp')

# Add the other three countries to 'world'
world = pd.concat([world, other], ignore_index=True)

# Finally, save to a shapefile
world.to_file('world_new\\world.shp')
